docs: add contribution guidelines
