+++ README.md
@@ -38,3 +38,7 @@ pytest -q
 - `config/` – 設定
 - `docs/` – ドキュメント
 - `tests/` – テスト
+
+## 貢献ガイド
+- コミットメッセージは命令形・現在形で72文字以内。
+- 変更後は `pytest -q` を実行してテストを確認してください。
